// Server/src/Types/Weather.ts
interface Weather {
  id: String;
  weatherState: String;
  weatherStateAbbr: String;
  humidity: String;
  windDirection: String;
  windSpeed: String;
  minTemp: String;
  maxTemp: String;
  theTemp: String;
  airPressure: String;
  visibility: String;
  predictability: String;
  icon: String;
}

export { Weather };